requirements.txt
flask
opencv-python
torch
torchvision
numpy
pillow
gunicorn
